Add holes on bottom for Arduino switching-board...
Move holes for 40-pin connectors on back-plate down 2mm, to align
properly.

Increase size of hole on back-plate for back of ATX power supply.  This
adds support for power supplies with the power cable socket at the top
_or_ bottom, when power supply is oriented with fan on the left when
looking from the back.

Update SVG files to match the SketchUp model.